ASL_Translator_Python/requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python
